Add a test requiring basic deferring to work in order to compile.
Currently, the first static if searching for T will trigger a second determineMembers call on XYZ. Then the second static if isn't able to find X, which won't be available until the first determineMembers call add the first static if member.

Another complexity of this example is that the first search for T can't be confident as long as determineMembers(XYZ) adds all the other members of XYZ (other than the first static if).